+++ misc/quiz.md
@@ -1,0 +1,45 @@
+Quiz: Wyjątki w Javie – poziom zaawansowany
+1. Która z poniższych klas NIE należy do hierarchii wyjątków w Javie?
+
+A. Throwable
+B. Exception
+C. RuntimeException
+D. String
+
+✅ Poprawna odpowiedź: D
+
+2. Czym charakteryzują się wyjątki typu checked?
+
+A. Są sprawdzane tylko w czasie działania programu
+B. Muszą być zadeklarowane lub obsłużone w kodzie
+C. Występują wyłącznie w systemach czasu rzeczywistego
+D. Nie mogą być dziedziczone
+
+✅ Poprawna odpowiedź: B
+
+3. Czym jest „exception chaining”?
+
+A. Tworzeniem pętli nieskończonego rzucania wyjątków
+B. Grupowaniem wyjątków tego samego typu
+C. Przekazywaniem jednego wyjątku jako przyczyny drugiego
+D. Obsługą wielu wyjątków w jednym bloku catch
+
+✅ Poprawna odpowiedź: C
+
+4. Które z poniższych stwierdzeń dotyczących własnych wyjątków w Javie jest prawdziwe?
+
+A. Można je tworzyć tylko jako unchecked
+B. Nie można ich dziedziczyć po Throwable
+C. Tworzy się je przez dziedziczenie po Exception lub RuntimeException
+D. Własne wyjątki są dostępne tylko w Javie 11+
+
+✅ Poprawna odpowiedź: C
+
+5. Która praktyka jest uznawana za dobrą przy obsłudze wyjątków?
+
+A. Rzucanie wyjątków do kontroli przepływu programu
+B. Wielokrotne logowanie tego samego wyjątku
+C. Zawsze używanie finally lub try-with-resources
+D. Ignorowanie wyjątków w blokach catch
+
+✅ Poprawna odpowiedź: C
