+++ misc/solves/zadanie2_solved.java
@@ -1,18 +1,15 @@
-import java.util.Scanner;
-import java.util.InputMismatchException;
-
-// Task: Add a finally block to close the Scanner
-
 public class FinallyBlock {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        ______ {
+        try {
+            System.out.print("Enter a number: ");
             int num = scanner.nextInt();
             System.out.println("Number: " + num);
-        } ______ (InputMismatchException e) {
+        } catch (InputMismatchException e) {
             System.out.println("Invalid input!");
-        } ______ {
+        } finally {
             System.out.println("Closing scanner...");
+            scanner.close();
         }
     }
 }
