+++ misc/solves/zadanie3_solved.java
@@ -1,0 +1,25 @@
+class InsufficientBalanceException extends Exception {
+    public InsufficientBalanceException(String message) {
+        super(message);
+    }
+}
+
+class BankAccount {
+    private double balance = 100.0;
+
+    public void withdraw(double amount) throws InsufficientBalanceException {
+        if (amount > balance) {
+            throw new InsufficientBalanceException("Balance cannot be negative!");
+        }
+        balance -= amount;
+    }
+
+    public static void main(String[] args) {
+        BankAccount account = new BankAccount();
+        try {
+            account.withdraw(150.0); // Test insufficient balance
+        } catch (InsufficientBalanceException e) {
+            System.out.println("Error: " + e.getMessage());
+        }
+    }
+}
