+++ misc/zadania.md
@@ -1,0 +1,20 @@
+## Zadanie 1
+
+**Polecenie:**  
+Uzupełnij obsługę błędów w pliku `MultiCatch.java`, stosując odpowiedni blok `catch` dla wyjątków klas ArithmeticException oraz NullPointerException.
+
+---
+
+## Zadanie 2
+
+**Polecenie:**  
+Uzupełnij blok `finally` oraz kod w pliku `FinallyBlock.java`, aby obiekt `scanner` został zamknięty.
+
+---
+
+## Zadanie 3
+
+**Polecenie:**  
+Zmodyfikuj kod w pliku `BankAccount.java`, aby w przypadku ujemnego salda konta zgłaszany był własny wyjątek użytkownika.
+
+---
